fix:  safety class so it can be updated by board members.  vice president report link does not work.  next month doesn't work on class calendar removed some logging statements
