+++ wpa_project/minutes/templates/minutes/minutes_form.html
@@ -85,7 +85,7 @@
                 Vice President's report
             </div>
             <div class="col">
-                <div id="div-vp">
+                <div id="div-vice">
                     <input type="hidden" id="vice-report-count" class="report-count" value="{{ reports.vice|length }}">
                     {% for report in reports.vice %}
                         {% include "minutes/forms/report_form.html" %}
