+++ wpa_project/program_app/temp_tests.py
@@ -1,11 +1,11 @@
-# from .tests.tests_beginner_class import TestsBeginnerClass
-# from .tests.tests_beginner_class import TestsBeginnerClass2
-# from .tests.tests_calendar import TestsCalendar
-# from .tests.tests_class_attendance import TestsClassAttendance
+from .tests.tests_beginner_class import TestsBeginnerClass
+from .tests.tests_beginner_class import TestsBeginnerClass2
+from .tests.tests_calendar import TestsCalendar
+from .tests.tests_class_attendance import TestsClassAttendance
 # from .tests.tests_class_payment import TestsClassPayment
 # from .tests.tests_class_registration import TestsClassRegistration
-# from .tests.tests_class_sign_in import TestsClassSignIn
-# from .tests.tests_signal import TestsSignal
+from .tests.tests_class_sign_in import TestsClassSignIn
+from .tests.tests_signal import TestsSignal
 # from .tests.tests_unregister import TestsUnregisterStudent
 # from .tests.tests_unregister import TestsUnregisterStudent2
 from .tests.tests_update_programs import TestsUpdatePrograms
