+++ wpa_project/program_app/views/class_calendar_view.py
@@ -4,7 +4,6 @@ from django.views.generic.base import View
 from django.utils import timezone
 from django.utils.safestring import mark_safe
 from django.http import HttpResponse
-from ..models import BeginnerClass
 from ..src import Calendar
 logger = logging.getLogger(__name__)
 
@@ -15,8 +14,11 @@ class ClassCalendarView(View):
         d = timezone.localtime(timezone.now()).date()
         if month is None:
             month = d.month
-        logging.debug(month)
 
+        if month > 12:
+            while month > 12:
+                d = d.replace(year=d.year + 1)
+                month = month - 12
         # Instantiate our calendar class with today's year and selected month
         cal = Calendar(d.year, month)
