+++ wpa_project/student_app/serializers/student_serializers.py
@@ -7,4 +7,4 @@ class StudentSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = Student
-        fields = ['id', 'first_name', 'last_name', 'dob', 'email']
+        fields = ['id', 'first_name', 'last_name', 'dob', 'email', 'safety_class']
