+++ wpa_project/student_app/temp_tests.py
@@ -1,5 +1,5 @@
 # from .tests.tests_index import TestsIndex
-from .tests.tests_instructor import TestsInstructor
+# from .tests.tests_instructor import TestsInstructor
 # from .tests.tests_other import TestsOther
 # from .tests.tests_search import TestsSearch
 # from .tests.tests_student import TestsStudent
